+++ files/wsl-setup.ps1
@@ -1,0 +1,116 @@
+# Install WSL
+
+
+
+function Update-Kernel () {
+    Write-Host(" ...Downloading WSL2 Kernel Update.")
+    $kernelURI = 'https://wslstorestorage.blob.core.windows.net/wslblob/wsl_update_x64.msi'
+    $kernelUpdate = ((Get-Location).Path) + '\wsl_update_x64.msi'
+    (New-Object System.Net.WebClient).DownloadFile($kernelURI, $kernelUpdate)
+    Write-Host(" ...Installing WSL2 Kernel Update.")
+    msiexec /i $kernelUpdate /qn
+    Start-Sleep -Seconds 5
+    Write-Host(" ...Cleaning up Kernel Update installer.")
+    Remove-Item -Path $kernelUpdate
+}
+
+function Get-Kernel-Updated () {
+    # Check for Kernel Update Package
+    Write-Host("Checking for Windows Subsystem for Linux Update...")
+    $uninstall64 = Get-ChildItem "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall" | ForEach-Object { Get-ItemProperty $_.PSPath } | Select-Object DisplayName, Publisher, DisplayVersion, InstallDate
+    if ($uninstall64.DisplayName -contains 'Windows Subsystem for Linux Update') {
+        return $true
+    } else {
+        return $false
+    }
+}
+
+
+if(!([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] 'Administrator')) {
+    "`n" + ('-' * 79)
+    "This script option requires privileges to activate Windows Features."
+    "This script will self elevate to run as an Administrator and continue"
+    "" + ('-' * 79)
+    $args
+    'Press any key to continue...'
+    $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
+
+    Start-Process -FilePath PowerShell.exe -Verb Runas -ArgumentList "-File `"$($MyInvocation.MyCommand.Path)`"  $($args)"
+    Exit
+   }
+
+Write-Host @'
+ =============================
+< Windows Subsystem for Linux >
+<        installer            >
+ =============================
+   \
+    \
+        .--.
+       |o_o |
+       |:_/ |
+      //   \ \
+     (|     | )
+    /'\_   _/`\
+    \___)=(___/
+'@
+
+
+Write-Host("Checking for Windows Version...")
+if ([Environment]::OSVersion.Version.Major -lt 10) {
+    Write-Error 'Upgrade to Windows 10 before running this script'
+    Exit
+}
+
+Write-Host("Checking for Windows Minimum Release...")
+if ((Get-ItemProperty "HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion").ReleaseId -lt 1803) {
+    Write-Error 'You need to run Windows Update and install Feature Updates to at least version 1803'
+    Exit
+}
+
+Write-Host("Checking for Windows Subsystem for Linux...")
+$rebootRequired = $false
+if ((Get-WindowsOptionalFeature -Online -FeatureName Microsoft-Windows-Subsystem-Linux).State -ne 'Enabled'){
+    Write-Host(" ...Installing Windows Subsystem for Linux.")
+    $wslinst = Enable-WindowsOptionalFeature -Online -NoRestart -FeatureName Microsoft-Windows-Subsystem-Linux
+    if ($wslinst.Restartneeded -eq $true){
+        $rebootRequired = $true
+    }
+} else {
+    Write-Host(" ...Windows Subsystem for Linux already installed.")
+}
+
+Write-Host("Checking for Virtual Machine Platform...")
+if ((Get-WindowsOptionalFeature -Online -FeatureName VirtualMachinePlatform).State -ne 'Enabled'){
+    Write-Host(" ...Installing Virtual Machine Platform.")
+    $vmpinst = Enable-WindowsOptionalFeature -Online -NoRestart -FeatureName VirtualMachinePlatform
+    if ($vmpinst.RestartNeeded -eq $true){
+        $rebootRequired = $true
+    }
+} else {
+    Write-Host(" ...Virtual Machine Platform already installed.")
+}
+
+
+if ($rebootRequired) {
+    shutdown /t 120 /r /c "Reboot required to finish installing WSL2"
+    $cancelReboot = Read-Host 'Cancel reboot for now (you still need to reboot and rerun to finish installing WSL2) [y/N]'
+    if ($cancelReboot.Length -ne 0){
+        if ($cancelReboot.Substring(0,1).ToLower() -eq 'y'){
+            shutdown /a
+        }
+    }
+} else {
+    if (!(Get-Kernel-Updated)) {
+        Write-Host(" ...WSL kernel update not installed.")
+        Update-Kernel
+    } else {
+        Write-Host(" ...WSL update already installed.")
+    }
+    Write-Host("Setting WSL2 as the default...")
+    wsl --set-default-version 2
+}
+
+Write-Host -NoNewLine 'Press any key to continue...';
+$null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
+